+++ hal_psee_plugins/include/devices/gen41/gen41_evk2_tz_device.h
@@ -41,7 +41,7 @@ public:
     virtual I_HW_Identification::SensorInfo get_sensor_info() {
         return {4, 1, "Gen4.1"};
     }
-    long long get_sensor_id();
+    int64_t get_sensor_id();
     virtual int get_illumination();
     virtual int get_temperature();
