+++ hal_psee_plugins/src/facilities/psee_device_control.cpp
@@ -82,7 +82,7 @@ I_CameraSynchronization::SyncMode PseeDeviceControl::get_mode() const {
     return sync_mode_;
 }
 
-long long PseeDeviceControl::get_sensor_id() {
+int64_t PseeDeviceControl::get_sensor_id() {
     return -1;
 }
